backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
openai==1.59.9
pydantic==2.10.5
pydantic-core==2.27.2
httpx==0.28.1
python-multipart==0.0.19